// src/app/interfaces/formInscription.ts
export interface FormInscription {
  idEmail: number;
  title: string;
  userName: string;
  userLastName: string;
  userEmail: string;
  userPhoneAreaCode: string;
  userPhone: number;
  userOtherPhoneAreaCode?: string;
  userOtherPhone?: number;
  userAddress: string;
  userAddress2?: string;
  userCity: string;
  userRegion: string;
  userPostalCode: number;
  companyName: string;
  CIF: string;
  companyEmail: string;
  companySector: string;
  companyPhoneAreaCode: string;
  companyPhone: number;
  companyOtherPhoneAreaCode?: string;
  companyOtherPhone?: number;
  companyContactPerson: string;
  companyYear: number;
  companyWorks: number;
  companyAddress: string;
  companyAddress2?: string;
  companyCity: string;
  companyRegion: string;
  companyPostalCode: number;
  companyDistrict: string;
  companyWorkArea: string;
  companyRedWeb: string;
  companyRedFacebook: string;
  companyRedInstagram: string;
  companyRedTwitter: string;
  companyRedLinkedin: string;
  companyRedTikTok: string;
  date: string;
  terms: boolean;
}